Move SelectOptions to new file
