+++ packages/components/src/Form/Inputs/Select/Select.tsx
@@ -24,24 +24,23 @@
 
  */
 
-import React, { forwardRef, Ref, FormEvent, ReactNode } from 'react'
+import React, { forwardRef, Ref, FormEvent } from 'react'
 import styled from 'styled-components'
 import { CustomizableAttributes } from '@looker/design-tokens'
-import { Box } from '../../../Layout'
-import { ListItem } from '../../../List'
-import { Heading, Paragraph } from '../../../Text'
 import { ValidationType } from '../../ValidationMessage'
 import {
   Combobox,
   ComboboxInput,
   ComboboxList,
-  ComboboxOption,
-  comboboxOptionGrid,
-  ComboboxOptionObject,
-  ComboboxOptionText,
   ComboboxProps,
   getComboboxText,
 } from '../Combobox'
+import {
+  SelectOptionGroupProps,
+  SelectOptionObject,
+  SelectOptionProps,
+  SelectOptions,
+} from './SelectOptions'
 
 export const CustomizableSelectAttributes: CustomizableAttributes = {
   borderRadius: 'medium',
@@ -50,17 +49,6 @@ export const CustomizableSelectAttributes: CustomizableAttributes = {
   px: 'xsmall',
   py: 'none',
 }
-
-export interface SelectOptionObject extends ComboboxOptionObject {
-  description?: string | ReactNode
-}
-
-export interface SelectOptionGroupProps {
-  options: SelectOptionObject[]
-  title: string | ReactNode
-}
-
-export type SelectOptionProps = SelectOptionObject | SelectOptionGroupProps
 
 export interface SelectBaseProps {
   options?: SelectOptionProps[]
@@ -127,57 +115,6 @@ function getFirstOption(options: SelectOptionProps[]): SelectOptionObject {
   if (optionAsGroup.title) return optionAsGroup.options[0]
   return options[0] as SelectOptionObject
 }
-
-export function SelectOptionWithDescription({
-  description,
-}: SelectOptionObject) {
-  return (
-    <Box>
-      <Heading fontSize="small" fontWeight="semiBold" pb="xxsmall">
-        <ComboboxOptionText />
-      </Heading>
-      <Paragraph variant="subdued" fontSize="small">
-        {description}
-      </Paragraph>
-    </Box>
-  )
-}
-
-const renderOption = (option: SelectOptionObject, index: number) => {
-  if (option.description) {
-    return (
-      <ComboboxOption {...option} key={index} py="xxsmall">
-        <SelectOptionWithDescription {...option} />
-      </ComboboxOption>
-    )
-  }
-  return <ComboboxOption {...option} key={index} />
-}
-
-const SelectOptionGroupTitle = styled(Heading)`
-  ${comboboxOptionGrid}
-`
-
-SelectOptionGroupTitle.defaultProps = {
-  fontSize: 'xxsmall',
-  fontWeight: 'semiBold',
-  px: 'xsmall',
-  py: 'xxsmall',
-  variant: 'subdued',
-}
-
-export const SelectOptionGroup = ({
-  options,
-  title,
-}: SelectOptionGroupProps) => (
-  <Box py="xxsmall">
-    <SelectOptionGroupTitle>
-      <span />
-      {title}
-    </SelectOptionGroupTitle>
-    {options.map(renderOption)}
-  </Box>
-)
 
 const SelectComponent = forwardRef(
   (
@@ -250,20 +187,7 @@ const SelectComponent = forwardRef(
         />
         {!disabled && (
           <ComboboxList persistSelection {...ariaProps}>
-            {options && options.length > 0 ? (
-              options.map((option: SelectOptionProps, index: number) => {
-                const optionAsGroup = option as SelectOptionGroupProps
-                return optionAsGroup.title ? (
-                  <SelectOptionGroup key={index} {...optionAsGroup} />
-                ) : (
-                  renderOption(option as SelectOptionObject, index)
-                )
-              })
-            ) : (
-              <ListItem fontSize="small" px="medium" py="xxsmall">
-                No options
-              </ListItem>
-            )}
+            <SelectOptions options={options} />
           </ComboboxList>
         )}
       </Combobox>
